// rotv-flight/src/gen/rotv/v1/telemetry_pb.d.ts
export declare type TelemetryStreamRequest = Message<"rotv.v1.TelemetryStreamRequest"> & {
  /**
   * @generated from field: string system_id = 1;
   */
  systemId: string;
};

/**
 * Describes the message rotv.v1.TelemetryStreamRequest.
 * Use `create(TelemetryStreamRequestSchema)` to create a new message.
 */
export declare const TelemetryStreamRequestSchema: GenMessage<TelemetryStreamRequest>;
// ...
export declare type TelemetrySnapshot = Message<"rotv.v1.TelemetrySnapshot"> & {
  /**
   * ISO 8601
   *
   * @generated from field: string timestamp = 1;
   */
  timestamp: string;

  /**
   * @generated from field: rotv.v1.Attitude attitude = 2;
   */
  attitude?: Attitude;

  /**
   * @generated from field: rotv.v1.Position position = 3;
   */
  position?: Position;

  /**
   * @generated from field: rotv.v1.Velocity velocity = 4;
   */
  velocity?: Velocity;

  /**
   * @generated from field: rotv.v1.Power power = 5;
   */
  power?: Power;

  /**
   * kN
   *
   * @generated from field: double tow_cable_tension = 6;
   */
  towCableTension: number;

  /**
   * °C
   *
   * @generated from field: double water_temp = 7;
   */
  waterTemp: number;

  /**
   * PSU
   *
   * @generated from field: double salinity = 8;
   */
  salinity: number;

  /**
   * m/s
   *
   * @generated from field: double sound_velocity = 9;
   */
  soundVelocity: number;

  /**
   * metres
   *
   * @generated from field: double cable_out = 10;
   */
  cableOut: number;
};

/**
 * Describes the message rotv.v1.TelemetrySnapshot.
 * Use `create(TelemetrySnapshotSchema)` to create a new message.
 */
export declare const TelemetrySnapshotSchema: GenMessage<TelemetrySnapshot>;

/**
 * @generated from service rotv.v1.TelemetryService
 */
export declare const TelemetryService: GenService<{
  /**
   * @generated from rpc rotv.v1.TelemetryService.StreamTelemetry
   */
  streamTelemetry: {
    methodKind: "server_streaming";
    input: typeof TelemetryStreamRequestSchema;
    output: typeof TelemetrySnapshotSchema;
  },
}>;